Updated/fixed tests
Combined new test for infer_file_type with existing one in submission_test
